+++ config/config.exs
@@ -1,7 +1,6 @@
 # This file is responsible for configuring your application
 # and its dependencies with the aid of the Mix.Config module.
-use Mix.Config
-
+import Config
 config :ecto_shorts, repo: nil, error_module: EctoShorts.Actions.Error
 
 if Mix.env() == :test do